+++ src/YnloGraphQLBundle.php
@@ -18,9 +18,9 @@ use Ynlo\GraphQLBundle\DependencyInjection\YnloGraphQLExtension;
 use Ynlo\GraphQLBundle\Encoder\IDEncoderManager;
 use Ynlo\GraphQLBundle\Error\Exporter\ErrorListExporterInterface;
 use Ynlo\GraphQLBundle\Extension\ExtensionInterface;
+use Ynlo\GraphQLBundle\Filter\FilterInterface;
 use Ynlo\GraphQLBundle\Filter\FilterResolverInterface;
 use Ynlo\GraphQLBundle\Form\Input\InputFieldTypeGuesser;
-use Ynlo\GraphQLBundle\OrderBy\OrderByInterface;
 use Ynlo\GraphQLBundle\Type\Loader\TypeAutoLoader;
 use Ynlo\GraphQLBundle\Type\Registry\TypeRegistry;
 use Ynlo\GraphQLBundle\Util\IDEncoder;
@@ -54,6 +54,9 @@ class YnloGraphQLBundle extends Bundle
         $container->registerForAutoconfiguration(ExtensionInterface::class)->addTag('graphql.extension');
         $container->registerForAutoconfiguration(InputFieldTypeGuesser::class)->addTag('graphql.input_type_guesser');
         $container->registerForAutoconfiguration(FilterResolverInterface::class)->addTag('graphql.filter_resolver');
+        $container->registerForAutoconfiguration(FilterInterface::class)
+                  ->addTag('graphql.list_filter')
+                  ->setPublic(true);
     }
 
     /**